% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boundaries.R
\name{boundaries}
\alias{boundaries}
\title{Boundaries for U.S. states, counties, or commuting zones}
\source{
U.S. Census Bureau. \href{https://www2.census.gov/geo/tiger/GENZ2018/shp/}{2018 Cartographic Boundary Shapefiles.}

USDA Economic Research Service.
\href{https://www.ers.usda.gov/data-products/commuting-zones-and-labor-market-areas/}{1990 Commuting Zones.}
}
\usage{
boundaries(
  geography = c("state", "county", "cz"),
  resolution = c("20m", "5m", "500k"),
  projection = c("albers", "longlat")
)
}
\arguments{
\item{geography}{One of:
\itemize{
\item \code{"state"} (Default)
\item \code{"county"}
\item \code{"cz"} Commuting zone.
}}

\item{resolution}{One of:
\itemize{
\item \code{"20m"} 1:20,000,000 (Default)
\item \code{"5m"} 1:5,000,000
\item \code{"500k"} 1:500,000
}}

\item{projection}{One of:
\itemize{
\item \code{"albers"} Albers equal area with Alaska and Hawaii scaled and below
contiguous states. (Default)
\item \code{"longlat"} Longitude and latitude.
}}
}
\value{
An sf object.
}
\description{
Boundaries for U.S. states, counties, or commuting zones as sf object.
}
\details{
The default 1:20,000,000 resolution is suitable for maps of the
entire U.S. The other two higher resolutions are suitable for maps of smaller
areas. The default Albers equal area projection is suitable for maps of the
entire U.S. The longitude and latitude projection is suitable for maps of
smaller areas.
}
\examples{
counties <- boundaries(geography = "county")

}
